+++ src/main/java/com/shootdoori/match/exception/common/BusinessException.java
@@ -1,4 +1,4 @@
-package com.shootdoori.match.exception;
+package com.shootdoori.match.exception.common;
 
 public class BusinessException extends RuntimeException {
