+++ src/main/java/com/shootdoori/match/exception/common/DuplicatedException.java
@@ -1,4 +1,4 @@
-package com.shootdoori.match.exception;
+package com.shootdoori.match.exception.common;
 
 public class DuplicatedException extends BusinessException {
