+++ src/main/java/com/shootdoori/match/exception/common/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.shootdoori.match.exception;
+package com.shootdoori.match.exception.common;
 
 public class NotFoundException extends BusinessException {
