+++ src/main/java/com/shootdoori/match/service/MatchCompleteService.java
@@ -5,8 +5,8 @@ import com.shootdoori.match.dto.TeamResponseDto;
 import com.shootdoori.match.entity.Match;
 import com.shootdoori.match.entity.Team;
 import com.shootdoori.match.entity.TeamMember;
-import com.shootdoori.match.exception.ErrorCode;
-import com.shootdoori.match.exception.NotFoundException;
+import com.shootdoori.match.exception.common.ErrorCode;
+import com.shootdoori.match.exception.common.NotFoundException;
 import com.shootdoori.match.repository.MatchRepository;
 import com.shootdoori.match.repository.TeamMemberRepository;
 import com.shootdoori.match.repository.TeamRepository;
