+++ src/main/java/com/shootdoori/match/service/MercenaryRecruitmentService.java
@@ -7,8 +7,8 @@ import com.shootdoori.match.entity.MercenaryRecruitment;
 import com.shootdoori.match.entity.MercenaryPosition;
 import com.shootdoori.match.entity.SkillLevel;
 import com.shootdoori.match.entity.Team;
-import com.shootdoori.match.exception.NotFoundException;
-import com.shootdoori.match.exception.ErrorCode;
+import com.shootdoori.match.exception.common.NotFoundException;
+import com.shootdoori.match.exception.common.ErrorCode;
 import com.shootdoori.match.repository.MercenaryRecruitmentRepository;
 import com.shootdoori.match.repository.TeamRepository;
 import org.springframework.data.domain.Page;
