+++ src/main/java/com/shootdoori/match/service/ProfileService.java
@@ -4,11 +4,10 @@ import com.shootdoori.match.dto.ProfileCreateRequest;
 import com.shootdoori.match.dto.ProfileMapper;
 import com.shootdoori.match.dto.ProfileResponse;
 import com.shootdoori.match.dto.ProfileUpdateRequest;
-import com.shootdoori.match.entity.TeamMember;
 import com.shootdoori.match.entity.User;
-import com.shootdoori.match.exception.DuplicatedException;
-import com.shootdoori.match.exception.ErrorCode;
-import com.shootdoori.match.exception.NotFoundException;
+import com.shootdoori.match.exception.common.DuplicatedException;
+import com.shootdoori.match.exception.common.ErrorCode;
+import com.shootdoori.match.exception.common.NotFoundException;
 import com.shootdoori.match.repository.ProfileRepository;
 import com.shootdoori.match.repository.RefreshTokenRepository;
 import com.shootdoori.match.repository.TeamMemberRepository;
