+++ src/main/java/com/shootdoori/match/util/JwtUtil.java
@@ -1,7 +1,7 @@
 package com.shootdoori.match.util;
 
 import com.shootdoori.match.entity.User;
-import com.shootdoori.match.exception.UnauthorizedException;
+import com.shootdoori.match.exception.common.UnauthorizedException;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.http.HttpServletRequest;
