+++ src/test/java/com/shootdoori/joinWaiting/JoinWaitingTest.java
@@ -11,8 +11,8 @@ import com.shootdoori.match.entity.TeamMember;
 import com.shootdoori.match.entity.TeamMemberRole;
 import com.shootdoori.match.entity.TeamType;
 import com.shootdoori.match.entity.User;
-import com.shootdoori.match.exception.JoinWaitingNotPendingException;
-import com.shootdoori.match.exception.NoPermissionException;
+import com.shootdoori.match.exception.domain.joinwaiting.JoinWaitingNotPendingException;
+import com.shootdoori.match.exception.common.NoPermissionException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
