+++ src/test/java/com/shootdoori/match/service/PasswordResetServiceTest.java
@@ -3,7 +3,7 @@ package com.shootdoori.match.service;
 import com.shootdoori.match.entity.PasswordOtpToken;
 import com.shootdoori.match.entity.PasswordResetToken;
 import com.shootdoori.match.entity.User;
-import com.shootdoori.match.exception.UnauthorizedException;
+import com.shootdoori.match.exception.common.UnauthorizedException;
 import com.shootdoori.match.repository.PasswordOtpTokenRepository;
 import com.shootdoori.match.repository.PasswordResetTokenRepository;
 import com.shootdoori.match.repository.ProfileRepository;
