+++ src/test/java/com/shootdoori/team/TeamTest.java
@@ -8,10 +8,10 @@ import com.shootdoori.match.entity.Team;
 import com.shootdoori.match.entity.TeamMemberRole;
 import com.shootdoori.match.entity.TeamType;
 import com.shootdoori.match.entity.User;
-import com.shootdoori.match.exception.DifferentException;
-import com.shootdoori.match.exception.DuplicatedException;
-import com.shootdoori.match.exception.LastTeamMemberRemovalNotAllowedException;
-import com.shootdoori.match.exception.TeamCapacityExceededException;
+import com.shootdoori.match.exception.common.DifferentException;
+import com.shootdoori.match.exception.common.DuplicatedException;
+import com.shootdoori.match.exception.domain.team.LastTeamMemberRemovalNotAllowedException;
+import com.shootdoori.match.exception.domain.team.TeamCapacityExceededException;
 import com.shootdoori.match.value.MemberCount;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
