+++ src/test/java/com/shootdoori/teamMember/TeamMemberServiceTest.java
@@ -16,10 +16,10 @@ import com.shootdoori.match.entity.TeamMember;
 import com.shootdoori.match.entity.TeamMemberRole;
 import com.shootdoori.match.entity.TeamType;
 import com.shootdoori.match.entity.User;
-import com.shootdoori.match.exception.DifferentException;
-import com.shootdoori.match.exception.DuplicatedException;
-import com.shootdoori.match.exception.NoPermissionException;
-import com.shootdoori.match.exception.NotFoundException;
+import com.shootdoori.match.exception.common.DifferentException;
+import com.shootdoori.match.exception.common.DuplicatedException;
+import com.shootdoori.match.exception.common.NoPermissionException;
+import com.shootdoori.match.exception.common.NotFoundException;
 import com.shootdoori.match.repository.ProfileRepository;
 import com.shootdoori.match.repository.TeamMemberRepository;
 import com.shootdoori.match.repository.TeamRepository;
